requirements.txt
# Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# PDF generation
reportlab>=3.6.0
Pillow>=9.0.0

# Configuration
python-dotenv>=0.19.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
pylint>=2.15.0
flake8>=5.0.0

# Logging and monitoring
structlog>=22.0.0
colorlog>=6.7.0

# Optional: Advanced features
# pandas>=1.5.0          # For data analysis
# matplotlib>=3.6.0      # For charts
# selenium>=4.5.0        # For dynamic content